fix(android): refactor cronet engine rebuild and cache deletion logic
